requirements.txt
flask==3.1.0
twilio==9.4.0
anthropic==0.42.0
sqlalchemy==2.0.36
apscheduler==3.10.4
python-dotenv==1.0.1
gunicorn==23.0.0
